requirements.txt
requests 
beautifulsoup4 
fake-useragent 
